Added the image_dir flag in the usage example
In most object detection workflows, image data is usually stored in an images directory with separate train and test folders. Explicitly wrote this using the --image_dir flag to indicate to the user that this is what the script assumes. This also indicates to the user that if the images are in a different directory, then it must be specified using this flag.

Also corrected the output path so that script writes in a data directory (as is the convention).